+++ examples/create/src/App.tsx
@@ -69,6 +69,7 @@ function App() {
 			onDragEnd={onDragEnd}
 			onResizeEnd={onResizeEnd}
 			onRangeChanged={setRange}
+			sidebarWidth={200}
 		>
 			<Timeline items={items} rows={rows} onCreateItem={onCreateItem} />
 		</TimelineContext>