+++ packages/dnd-timeline/src/types/timeline.ts
@@ -68,8 +68,6 @@ export interface TimelineBag {
 	timelineRef: React.MutableRefObject<HTMLElement | null>;
 	setTimelineRef: (element: HTMLElement | null) => void;
 	sidebarWidth: number;
-	sidebarRef: React.MutableRefObject<HTMLElement | null>;
-	setSidebarRef: (element: HTMLElement | null) => void;
 	valueToPixels: ValueToPixels;
 	pixelsToValue: PixelsToValue;
 	getValueFromScreenX: GetValueFromScreenX;
@@ -108,6 +106,7 @@ export interface UseTimelineProps {
 	usePanStrategy?: UsePanStrategy;
 	onRangeChanged: OnRangeChanged;
 	rangeGridSizeDefinition?: number | GridSizeDefinition[];
+	sidebarWidth: number;
 }
 
 export interface TimelineContextProps